treewide: use unsigned numbers for %o
%o doesn't handle negative numbers, but we currently use it to format
unsigned ints.  Avoid that by casting the numbers when we know they
are positive, or passin unsigned constants as appropriate.

This casting/passing unsigned numbers isn't strictly necessasy in any
of the places this patch changes, however it stops the compiler from
complaining when -Wformat-signedness is passed, so it helps turning it
on in a further step.

+++ builtin/init-db.c
@@ -394,7 +394,7 @@ int init_db(const char *git_dir, const char *real_git_dir,
 		 */
 		if (get_shared_repository() < 0)
 			/* force to the mode value */
-			xsnprintf(buf, sizeof(buf), "0%o", -get_shared_repository());
+			xsnprintf(buf, sizeof(buf), "0%o", (unsigned int)-get_shared_repository());
 		else if (get_shared_repository() == PERM_GROUP)
 			xsnprintf(buf, sizeof(buf), "%d", OLD_PERM_GROUP);
 		else if (get_shared_repository() == PERM_EVERYBODY)
